made the board lighter, and fixed some errors (still errors when there's a tie or a game over, but that's whatever

+++ Assets/Scripts/TileController.cs
@@ -9,10 +9,10 @@ public class TileController : MonoBehaviour
     public TileColor color;
     public Color baseColor;
     public Color highlightColor;
-    public int colorScale;
+    public float colorScale;
 
     SpriteRenderer spriteRenderer;
-    int highlightScale;
+    float highlightScale;
     // Start is called before the first frame update
     void Start()
     {
